+++ server/schemas/index.js
@@ -1,4 +1,4 @@
 const typeDefs = require('./typeDefs');
 const resolvers = require('./resolvers');
 
-module.exports = { typeDefs, resolvers }; //! Added resolvers to export
+module.exports = { typeDefs, resolvers };
