+++ server/server.js
@@ -1,12 +1,11 @@
 const express = require("express");
 const { ApolloServer } = require("apollo-server-express");
 const path = require("path");
-const { User, Review, Movie } = require("./models")
 
 // const routes = require("./routes"); //! We do not need this
 const { authMiddleware } = require("./utils/auth");
 
-const { typeDefs, resolvers } = require("./schema");
+const { typeDefs, resolvers } = require("./schemas");
 const db = require("./config/connection");
 
 const PORT = process.env.PORT || 3001;
@@ -15,13 +14,6 @@ const server = new ApolloServer({
   typeDefs,
   resolvers,
   context: authMiddleware,
-  dataSources: () => { //!Data sources for queries and mutations in resolvers
-    return {           
-      User,            
-      Review,
-      Movie
-    };
-  },
 });
 
 app.use(express.urlencoded({ extended: false }));
