+++ server/utils/auth.js
@@ -6,6 +6,7 @@ const expiration = "2h";
 module.exports = {
   authMiddleware: function ({ req }) {
     let token = req.body.token || req.query.token || req.headers.authorization;
+
     if (req.headers.authorization) {
       token = token.split(" ").pop().trim();
     }
@@ -17,7 +18,7 @@ module.exports = {
     try {
       const { data } = jwt.verify(token, secret, { maxAge: expiration });
       req.user = data;
-    } catch (err) {
+    } catch {
       console.log("Invalid token");
     }
